Move usePostMeta hook to /hooks directory
